encoder for avro
